move advertiser,detector to its own packages
